feat(datamodel): add SSVC attributes to Vulnerability on Event (#17565)
