+++ opencti-platform/opencti-front/lang/back/de.json
@@ -982,6 +982,9 @@
   "SSL verify": "SSL-Verifizierung",
   "SSO Advanced configuration": "SSO Erweiterte Konfiguration",
   "SSO Configuration": "SSO-Konfiguration",
+  "SSVC Automatable": "SSVC Automatisierbar",
+  "SSVC Exploitation": "SSVC Ausnutzung",
+  "SSVC Technical Impact": "SSVC Technische Auswirkung",
   "Standard id": "Standard-ID",
   "Start": "Start",
   "Start date": "Startdatum",
