+++ opencti-platform/opencti-front/lang/back/en.json
@@ -982,6 +982,9 @@
   "SSL verify": "SSL verify",
   "SSO Advanced configuration": "SSO Advanced configuration",
   "SSO Configuration": "SSO Configuration",
+  "SSVC Automatable": "SSVC Automatable",
+  "SSVC Exploitation": "SSVC Exploitation",
+  "SSVC Technical Impact": "SSVC Technical Impact",
   "Standard id": "Standard id",
   "Start": "Start",
   "Start date": "Start date",
