+++ opencti-platform/opencti-front/lang/back/es.json
@@ -982,6 +982,9 @@
   "SSL verify": "Verificación SSL",
   "SSO Advanced configuration": "SSO Configuración avanzada",
   "SSO Configuration": "Configuración SSO",
+  "SSVC Automatable": "SSVC Automatizable",
+  "SSVC Exploitation": "SSVC Explotación",
+  "SSVC Technical Impact": "SSVC Impacto técnico",
   "Standard id": "Id estándar",
   "Start": "Inicio",
   "Start date": "Fecha de inicio",
