+++ opencti-platform/opencti-front/lang/back/fr.json
@@ -982,6 +982,9 @@
   "SSL verify": "SSL verify",
   "SSO Advanced configuration": "Configuration avancée du SSO",
   "SSO Configuration": "Configuration SSO",
+  "SSVC Automatable": "SSVC Automatisable",
+  "SSVC Exploitation": "SSVC Exploitation",
+  "SSVC Technical Impact": "SSVC Impact technique",
   "Standard id": "Standard id",
   "Start": "Démarrage",
   "Start date": "Date de début",
