+++ opencti-platform/opencti-front/lang/back/it.json
@@ -982,6 +982,9 @@
   "SSL verify": "Verifica SSL",
   "SSO Advanced configuration": "SSO Configurazione avanzata",
   "SSO Configuration": "Configurazione SSO",
+  "SSVC Automatable": "SSVC Automatizzabile",
+  "SSVC Exploitation": "SSVC Sfruttamento",
+  "SSVC Technical Impact": "SSVC Impatto tecnico",
   "Standard id": "ID standard",
   "Start": "Inizio",
   "Start date": "Data di inizio",
