+++ opencti-platform/opencti-front/lang/back/ja.json
@@ -982,6 +982,9 @@
   "SSL verify": "SSLベリファイ",
   "SSO Advanced configuration": "SSO詳細設定",
   "SSO Configuration": "SSOの設定",
+  "SSVC Automatable": "SSVC 自動化可能性",
+  "SSVC Exploitation": "SSVC 悪用状況",
+  "SSVC Technical Impact": "SSVC 技術的影響",
   "Standard id": "標準ID",
   "Start": "期限",
   "Start date": "開始日",
