+++ opencti-platform/opencti-front/lang/back/ko.json
@@ -982,6 +982,9 @@
   "SSL verify": "SSL 확인",
   "SSO Advanced configuration": "SSO 고급 구성",
   "SSO Configuration": "SSO 구성",
+  "SSVC Automatable": "SSVC 자동화 가능성",
+  "SSVC Exploitation": "SSVC 악용",
+  "SSVC Technical Impact": "SSVC 기술적 영향",
   "Standard id": "표준 ID",
   "Start": "시작",
   "Start date": "시작 날짜",
