+++ opencti-platform/opencti-front/lang/back/ru.json
@@ -982,6 +982,9 @@
   "SSL verify": "Проверка SSL",
   "SSO Advanced configuration": "Расширенная конфигурация SSO",
   "SSO Configuration": "Конфигурация SSO",
+  "SSVC Automatable": "SSVC Автоматизируемость",
+  "SSVC Exploitation": "SSVC Эксплуатация",
+  "SSVC Technical Impact": "SSVC Техническое воздействие",
   "Standard id": "Стандартный идентификатор",
   "Start": "Начало",
   "Start date": "Дата начала",
