+++ opencti-platform/opencti-front/lang/back/zh.json
@@ -982,6 +982,9 @@
   "SSL verify": "SSL 校验",
   "SSO Advanced configuration": "SSO 高级配置",
   "SSO Configuration": "SSO 配置",
+  "SSVC Automatable": "SSVC 可自动化",
+  "SSVC Exploitation": "SSVC 利用情况",
+  "SSVC Technical Impact": "SSVC 技术影响",
   "Standard id": "标准 id",
   "Start": "开始",
   "Start date": "开始日期",
