+++ opencti-platform/opencti-front/lang/front/de.json
@@ -4430,6 +4430,9 @@
   "SS Integrity Impact (SI)": "SS Auswirkungen auf die Integrität (SI)",
   "SSO": "SSO",
   "SSO Binding type": "SSO-Bindungstyp",
+  "SSVC Automatable": "SSVC Automatisierbar",
+  "SSVC Exploitation": "SSVC Ausnutzung",
+  "SSVC Technical Impact": "SSVC Technische Auswirkung",
   "Stable": "Stabil",
   "Stable score": "Stabile Punktestand",
   "Stacked": "Gestapelt",
