+++ opencti-platform/opencti-front/lang/front/en.json
@@ -4430,6 +4430,9 @@
   "SS Integrity Impact (SI)": "SS Integrity Impact (SI)",
   "SSO": "SSO",
   "SSO Binding type": "SSO Binding type",
+  "SSVC Automatable": "SSVC Automatable",
+  "SSVC Exploitation": "SSVC Exploitation",
+  "SSVC Technical Impact": "SSVC Technical Impact",
   "Stable": "Stable",
   "Stable score": "Stable score",
   "Stacked": "Stacked",
