+++ opencti-platform/opencti-front/lang/front/es.json
@@ -4430,6 +4430,9 @@
   "SS Integrity Impact (SI)": "SS Impacto en la integridad (SI)",
   "SSO": "SSO",
   "SSO Binding type": "Tipo de enlace SSO",
+  "SSVC Automatable": "SSVC Automatizable",
+  "SSVC Exploitation": "SSVC Explotación",
+  "SSVC Technical Impact": "SSVC Impacto técnico",
   "Stable": "Estable",
   "Stable score": "Puntuación estable",
   "Stacked": "Apilado",
