+++ opencti-platform/opencti-front/lang/front/fr.json
@@ -4430,6 +4430,9 @@
   "SS Integrity Impact (SI)": "SS Impact sur l'intégrité (SI)",
   "SSO": "SSO",
   "SSO Binding type": "Type de liaison SSO",
+  "SSVC Automatable": "SSVC Automatisable",
+  "SSVC Exploitation": "SSVC Exploitation",
+  "SSVC Technical Impact": "SSVC Impact technique",
   "Stable": "Stable",
   "Stable score": "Score stable",
   "Stacked": "Empilé",
