+++ opencti-platform/opencti-front/lang/front/it.json
@@ -4430,6 +4430,9 @@
   "SS Integrity Impact (SI)": "SS Impatto sull'integrità (SI)",
   "SSO": "SSO",
   "SSO Binding type": "Tipo di vincolo SSO",
+  "SSVC Automatable": "SSVC Automatizzabile",
+  "SSVC Exploitation": "SSVC Sfruttamento",
+  "SSVC Technical Impact": "SSVC Impatto tecnico",
   "Stable": "Stabile",
   "Stable score": "Punteggio stabile",
   "Stacked": "Impilato",
