+++ opencti-platform/opencti-front/lang/front/ja.json
@@ -4430,6 +4430,9 @@
   "SS Integrity Impact (SI)": "SS 完全性への影響(SI)",
   "SSO": "SSO",
   "SSO Binding type": "SSO バインディングタイプ",
+  "SSVC Automatable": "SSVC 自動化可能性",
+  "SSVC Exploitation": "SSVC 悪用状況",
+  "SSVC Technical Impact": "SSVC 技術的影響",
   "Stable": "安定",
   "Stable score": "安定したスコア",
   "Stacked": "積み重ね",
