+++ opencti-platform/opencti-front/lang/front/ko.json
@@ -4430,6 +4430,9 @@
   "SS Integrity Impact (SI)": "SS 무결성 영향(SI)",
   "SSO": "SSO",
   "SSO Binding type": "SSO 바인딩 유형",
+  "SSVC Automatable": "SSVC 자동화 가능성",
+  "SSVC Exploitation": "SSVC 악용",
+  "SSVC Technical Impact": "SSVC 기술적 영향",
   "Stable": "Stable",
   "Stable score": "안정된 점수",
   "Stacked": "누적",
