+++ opencti-platform/opencti-front/lang/front/ru.json
@@ -4430,6 +4430,9 @@
   "SS Integrity Impact (SI)": "SS Integrity Impact (SI)",
   "SSO": "SSO",
   "SSO Binding type": "Тип привязки SSO",
+  "SSVC Automatable": "SSVC Автоматизируемость",
+  "SSVC Exploitation": "SSVC Эксплуатация",
+  "SSVC Technical Impact": "SSVC Техническое воздействие",
   "Stable": "Стабильный",
   "Stable score": "Стабильная оценка",
   "Stacked": "Штабелированный",
