+++ opencti-platform/opencti-front/lang/front/zh.json
@@ -4430,6 +4430,9 @@
   "SS Integrity Impact (SI)": "SS 完整性影响 (SI)",
   "SSO": "SSO",
   "SSO Binding type": "SSO 绑定类型",
+  "SSVC Automatable": "SSVC 可自动化",
+  "SSVC Exploitation": "SSVC 利用情况",
+  "SSVC Technical Impact": "SSVC 技术影响",
   "Stable": "稳定",
   "Stable score": "成绩稳定",
   "Stacked": "堆叠",
