+++ opencti-platform/opencti-graphql/src/modules/vulnerability/vulnerability-converter.ts
@@ -2,12 +2,10 @@ import { buildStixDomain } from '../../database/stix-2-1-converter';
 import { buildStixDomain as buildStixDomain_2_0 } from '../../database/stix-2-0-converter';
 import { assertType, cleanObject, convertToStixDate } from '../../database/stix-converter-utils';
 import { STIX_EXT_OCTI } from '../../types/stix-2-1-extensions';
-import type { StoreEntityVulnerability } from './vulnerability-types';
-import type * as SDO from '../../types/stix-2-1-sdo';
-import type * as SDO_2_0 from '../../types/stix-2-0-sdo';
+import type { StoreEntityVulnerability, StixVulnerability, Stix2Vulnerability } from './vulnerability-types';
 import { ENTITY_TYPE_VULNERABILITY } from './vulnerability-types';
 
-const convertVulnerabilityToStix = (instance: StoreEntityVulnerability): SDO.StixVulnerability => {
+const convertVulnerabilityToStix = (instance: StoreEntityVulnerability): StixVulnerability => {
   assertType(ENTITY_TYPE_VULNERABILITY, instance.entity_type);
   const vulnerability = buildStixDomain(instance);
   return {
@@ -62,6 +60,10 @@ const convertVulnerabilityToStix = (instance: StoreEntityVulnerability): SDO.Sti
         cvss_v4_availability_impact_v: instance.x_opencti_cvss_v4_availability_impact_v,
         cvss_v4_availability_impact_s: instance.x_opencti_cvss_v4_availability_impact_s,
         cvss_v4_exploit_maturity: instance.x_opencti_cvss_v4_exploit_maturity,
+        // SSVC
+        ssvc_exploitation: instance.x_opencti_ssvc_exploitation,
+        ssvc_automatable: instance.x_opencti_ssvc_automatable,
+        ssvc_technical_impact: instance.x_opencti_ssvc_technical_impact,
         // Others
         cwe: instance.x_opencti_cwe,
         cisa_kev: instance.x_opencti_cisa_kev,
@@ -74,7 +76,7 @@ const convertVulnerabilityToStix = (instance: StoreEntityVulnerability): SDO.Sti
   };
 };
 
-export const convertVulnerabilityToStix_2_0 = (instance: StoreEntityVulnerability): SDO_2_0.StixVulnerability => {
+export const convertVulnerabilityToStix_2_0 = (instance: StoreEntityVulnerability): Stix2Vulnerability => {
   assertType(ENTITY_TYPE_VULNERABILITY, instance.entity_type);
   const vulnerability = buildStixDomain_2_0(instance);
   return {
@@ -128,6 +130,10 @@ export const convertVulnerabilityToStix_2_0 = (instance: StoreEntityVulnerabilit
     x_opencti_cvss_v4_availability_impact_v: instance.x_opencti_cvss_v4_availability_impact_v,
     x_opencti_cvss_v4_availability_impact_s: instance.x_opencti_cvss_v4_availability_impact_s,
     x_opencti_cvss_v4_exploit_maturity: instance.x_opencti_cvss_v4_exploit_maturity,
+    // SSVC
+    x_opencti_ssvc_exploitation: instance.x_opencti_ssvc_exploitation,
+    x_opencti_ssvc_automatable: instance.x_opencti_ssvc_automatable,
+    x_opencti_ssvc_technical_impact: instance.x_opencti_ssvc_technical_impact,
     // Others
     x_opencti_score: instance.x_opencti_score,
     x_opencti_epss_score: instance.x_opencti_epss_score,
