+++ opencti-platform/opencti-graphql/src/modules/vulnerability/vulnerability-types.ts
@@ -1,9 +1,14 @@
 import type { BasicStoreEntity, StoreEntity } from '../../types/store';
+import type { StixDate, StixDomainObject, StixOpenctiExtension } from '../../types/stix-2-1-common';
+import type { StixDomainObject as StixDomainObject2 } from '../../types/stix-2-0-common';
+import { STIX_EXT_OCTI } from '../../types/stix-2-1-extensions';
 
 export const ENTITY_TYPE_VULNERABILITY = 'Vulnerability';
 
 export const CVSS_SEVERITY_VALUES = ['CRITICAL', 'HIGH', 'MEDIUM', 'LOW', 'Unknown'];
-
+export const SSVC_EXPLOITATION_VALUES = ['none', 'poc', 'active'];
+export const SSVC_AUTOMATABLE_VALUES = ['no', 'yes'];
+export const SSVC_TECHNICAL_IMPACT_VALUES = ['partial', 'total'];
 // region Database types
 export interface BasicStoreEntityVulnerability extends BasicStoreEntity {
   name: string;
@@ -53,6 +58,10 @@ export interface BasicStoreEntityVulnerability extends BasicStoreEntity {
   x_opencti_cvss_v4_availability_impact_v: string;
   x_opencti_cvss_v4_availability_impact_s: string;
   x_opencti_cvss_v4_exploit_maturity: string;
+  // SSVC
+  x_opencti_ssvc_exploitation: string;
+  x_opencti_ssvc_automatable: string;
+  x_opencti_ssvc_technical_impact: string;
   // Others
   x_opencti_cwe: Array<string>;
   x_opencti_cisa_kev: boolean;
@@ -63,4 +72,136 @@ export interface BasicStoreEntityVulnerability extends BasicStoreEntity {
 }
 
 export interface StoreEntityVulnerability extends BasicStoreEntityVulnerability, StoreEntity {}
+// endregion
+
+// region Stix 2.1 type
+export interface StixVulnerabilityExtension extends StixOpenctiExtension {
+  // CVSS3
+  cvss_vector: string;
+  cvss_base_score: number;
+  cvss_base_severity: string;
+  cvss_attack_vector: string;
+  cvss_attack_complexity: string;
+  cvss_privileges_required: string;
+  cvss_user_interaction: string;
+  cvss_scope: string;
+  cvss_confidentiality_impact: string;
+  cvss_integrity_impact: string;
+  cvss_availability_impact: string;
+  cvss_exploit_code_maturity: string;
+  cvss_remediation_level: string;
+  cvss_report_confidence: string;
+  cvss_temporal_score: number;
+  // CVSS2
+  cvss_v2_vector: string;
+  cvss_v2_base_score: number;
+  cvss_v2_access_vector: string;
+  cvss_v2_access_complexity: string;
+  cvss_v2_authentication: string;
+  cvss_v2_confidentiality_impact: string;
+  cvss_v2_integrity_impact: string;
+  cvss_v2_availability_impact: string;
+  cvss_v2_exploitability: string;
+  cvss_v2_remediation_level: string;
+  cvss_v2_report_confidence: string;
+  cvss_v2_temporal_score: number;
+  // CVSS4
+  cvss_v4_vector: string;
+  cvss_v4_base_score: number;
+  cvss_v4_base_severity: string;
+  cvss_v4_attack_vector: string;
+  cvss_v4_attack_complexity: string;
+  cvss_v4_attack_requirements: string;
+  cvss_v4_privileges_required: string;
+  cvss_v4_user_interaction: string;
+  cvss_v4_confidentiality_impact_v: string;
+  cvss_v4_confidentiality_impact_s: string;
+  cvss_v4_integrity_impact_v: string;
+  cvss_v4_integrity_impact_s: string;
+  cvss_v4_availability_impact_v: string;
+  cvss_v4_availability_impact_s: string;
+  cvss_v4_exploit_maturity: string;
+  // SSVC
+  ssvc_exploitation: string;
+  ssvc_automatable: string;
+  ssvc_technical_impact: string;
+  // Others
+  cwe: Array<string>;
+  cisa_kev: boolean;
+  score: number;
+  epss_score: number;
+  epss_percentile: number;
+  first_seen_active: StixDate;
+}
+
+// name, description
+export interface StixVulnerability extends StixDomainObject {
+  name: string;
+  description: string; // optional
+  extensions: {
+    [STIX_EXT_OCTI]: StixVulnerabilityExtension;
+  };
+}
+// endregion
+
+// region Stix 2.0 type
+export interface Stix2Vulnerability extends StixDomainObject2 {
+  name: string;
+  description: string; // optional
+  x_opencti_cisa_kev: boolean;
+  x_opencti_first_seen_active: Date;
+  // CVSS3
+  x_opencti_cvss_vector_string: string;
+  x_opencti_cvss_base_score: number;
+  x_opencti_cvss_base_severity: string;
+  x_opencti_cvss_attack_vector: string;
+  x_opencti_cvss_attack_complexity: string;
+  x_opencti_cvss_privileges_required: string;
+  x_opencti_cvss_user_interaction: string;
+  x_opencti_cvss_scope: string;
+  x_opencti_cvss_confidentiality_impact: string;
+  x_opencti_cvss_integrity_impact: string;
+  x_opencti_cvss_availability_impact: string;
+  x_opencti_cvss_exploit_code_maturity: string;
+  x_opencti_cvss_remediation_level: string;
+  x_opencti_cvss_report_confidence: string;
+  x_opencti_cvss_temporal_score: number;
+  // CVSS2
+  x_opencti_cvss_v2_vector_string: string;
+  x_opencti_cvss_v2_base_score: number;
+  x_opencti_cvss_v2_access_vector: string;
+  x_opencti_cvss_v2_access_complexity: string;
+  x_opencti_cvss_v2_authentication: string;
+  x_opencti_cvss_v2_confidentiality_impact: string;
+  x_opencti_cvss_v2_integrity_impact: string;
+  x_opencti_cvss_v2_availability_impact: string;
+  x_opencti_cvss_v2_exploitability: string;
+  x_opencti_cvss_v2_remediation_level: string;
+  x_opencti_cvss_v2_report_confidence: string;
+  x_opencti_cvss_v2_temporal_score: number;
+  // CVSS4
+  x_opencti_cvss_v4_vector_string: string;
+  x_opencti_cvss_v4_base_score: number;
+  x_opencti_cvss_v4_base_severity: string;
+  x_opencti_cvss_v4_attack_vector: string;
+  x_opencti_cvss_v4_attack_complexity: string;
+  x_opencti_cvss_v4_attack_requirements: string;
+  x_opencti_cvss_v4_privileges_required: string;
+  x_opencti_cvss_v4_user_interaction: string;
+  x_opencti_cvss_v4_confidentiality_impact_v: string;
+  x_opencti_cvss_v4_confidentiality_impact_s: string;
+  x_opencti_cvss_v4_integrity_impact_v: string;
+  x_opencti_cvss_v4_integrity_impact_s: string;
+  x_opencti_cvss_v4_availability_impact_v: string;
+  x_opencti_cvss_v4_availability_impact_s: string;
+  x_opencti_cvss_v4_exploit_maturity: string;
+  // SSVC
+  x_opencti_ssvc_exploitation: string;
+  x_opencti_ssvc_automatable: string;
+  x_opencti_ssvc_technical_impact: string;
+  // Others
+  x_opencti_score: number;
+  x_opencti_epss_score: number;
+  x_opencti_epss_percentile: number;
+}
 // endregion
