+++ opencti-platform/opencti-graphql/src/modules/vulnerability/vulnerability.graphql
@@ -20,6 +20,21 @@ type VulnerabilityEdge {
   cursor: String!
   node: Vulnerability!
 }
+enum SsvcExploitation {
+  none
+  poc
+  active
+}
+
+enum SsvcAutomatable {
+  no
+  yes
+}
+
+enum SsvcTechnicalImpact {
+  partial
+  total
+}
 type Vulnerability implements BasicObject & StixObject & StixCoreObject & StixDomainObject {
   id: ID! # internal_id
   standard_id: String!
@@ -181,6 +196,10 @@ type Vulnerability implements BasicObject & StixObject & StixCoreObject & StixDo
   x_opencti_cvss_v4_availability_impact_v: String
   x_opencti_cvss_v4_availability_impact_s: String
   x_opencti_cvss_v4_exploit_maturity: String
+  # SSVC
+  x_opencti_ssvc_exploitation: SsvcExploitation @ff(flags: ["SSVC_ATTRIBUTES"])
+  x_opencti_ssvc_automatable: SsvcAutomatable @ff(flags: ["SSVC_ATTRIBUTES"])
+  x_opencti_ssvc_technical_impact: SsvcTechnicalImpact @ff(flags: ["SSVC_ATTRIBUTES"])
   # Others
   x_opencti_cwe: [String]
   x_opencti_cisa_kev: Boolean
@@ -269,6 +288,10 @@ input VulnerabilityAddInput {
   x_opencti_cvss_v4_availability_impact_v: String
   x_opencti_cvss_v4_availability_impact_s: String
   x_opencti_cvss_v4_exploit_maturity: String
+  # SSVC
+  x_opencti_ssvc_exploitation: String
+  x_opencti_ssvc_automatable: String
+  x_opencti_ssvc_technical_impact: String
   # Others
   x_opencti_cwe: [String]
   x_opencti_cisa_kev: Boolean
