+++ opencti-platform/opencti-graphql/src/modules/vulnerability/vulnerability.ts
@@ -2,10 +2,17 @@ import { type ModuleDefinition, registerDefinition } from '../../schema/module';
 import { ABSTRACT_STIX_DOMAIN_OBJECT } from '../../schema/general';
 import { NAME_FIELD, normalizeName } from '../../schema/identifier';
 import { objectOrganization } from '../../schema/stixRefRelationship';
-import type * as SDO from '../../types/stix-2-1-sdo';
 import type { OverviewLayoutCustomization } from '../entitySetting/entitySetting-types';
 import { registerEntityOverviewLayoutCustomization } from '../../schema/overviewLayoutCustomization-register';
-import { CVSS_SEVERITY_VALUES, ENTITY_TYPE_VULNERABILITY, type StoreEntityVulnerability } from './vulnerability-types';
+import {
+  CVSS_SEVERITY_VALUES,
+  ENTITY_TYPE_VULNERABILITY,
+  SSVC_AUTOMATABLE_VALUES,
+  SSVC_EXPLOITATION_VALUES,
+  SSVC_TECHNICAL_IMPACT_VALUES,
+  type StixVulnerability,
+  type StoreEntityVulnerability,
+} from './vulnerability-types';
 import convertVulnerabilityToStix, { convertVulnerabilityToStix_2_0 } from './vulnerability-converter';
 
 const vulnerabilityDefaultOverviewLayout: OverviewLayoutCustomization[] = [
@@ -24,7 +31,7 @@ registerEntityOverviewLayoutCustomization(
   vulnerabilityDefaultOverviewLayout,
 );
 
-const VULNERABILITY_DEFINITION: ModuleDefinition<StoreEntityVulnerability, SDO.StixVulnerability> = {
+const VULNERABILITY_DEFINITION: ModuleDefinition<StoreEntityVulnerability, StixVulnerability> = {
   type: {
     id: 'vulnerability',
     name: ENTITY_TYPE_VULNERABILITY,
@@ -89,6 +96,10 @@ const VULNERABILITY_DEFINITION: ModuleDefinition<StoreEntityVulnerability, SDO.S
     { name: 'x_opencti_cvss_v4_availability_impact_v', label: 'CVSS4 VS Availability Impact (VA)', type: 'string', format: 'short', mandatoryType: 'no', editDefault: false, multiple: false, upsert: true, isFilterable: true },
     { name: 'x_opencti_cvss_v4_availability_impact_s', label: 'CVSS4 SS Availability Impact (SA)', type: 'string', format: 'short', mandatoryType: 'no', editDefault: false, multiple: false, upsert: true, isFilterable: true },
     { name: 'x_opencti_cvss_v4_exploit_maturity', label: 'CVSS4 Exploit Maturity (E)', type: 'string', format: 'short', mandatoryType: 'no', editDefault: false, multiple: false, upsert: true, isFilterable: true },
+    // SSVC
+    { name: 'x_opencti_ssvc_exploitation', label: 'SSVC Exploitation', type: 'string', format: 'enum', values: SSVC_EXPLOITATION_VALUES, mandatoryType: 'no', editDefault: false, multiple: false, upsert: true, isFilterable: true, featureFlag: 'SSVC_ATTRIBUTES' },
+    { name: 'x_opencti_ssvc_automatable', label: 'SSVC Automatable', type: 'string', format: 'enum', values: SSVC_AUTOMATABLE_VALUES, mandatoryType: 'no', editDefault: false, multiple: false, upsert: true, isFilterable: true, featureFlag: 'SSVC_ATTRIBUTES' },
+    { name: 'x_opencti_ssvc_technical_impact', label: 'SSVC Technical Impact', type: 'string', format: 'enum', values: SSVC_TECHNICAL_IMPACT_VALUES, mandatoryType: 'no', editDefault: false, multiple: false, upsert: true, isFilterable: true, featureFlag: 'SSVC_ATTRIBUTES' },
     // Others
     { name: 'x_opencti_cwe', label: 'Associated CWE(s)', type: 'string', format: 'short', mandatoryType: 'no', editDefault: false, multiple: true, upsert: true, isFilterable: true },
     { name: 'x_opencti_cisa_kev', label: 'CISA KEV', type: 'boolean', mandatoryType: 'no', editDefault: false, multiple: false, upsert: true, isFilterable: true },
@@ -99,7 +110,7 @@ const VULNERABILITY_DEFINITION: ModuleDefinition<StoreEntityVulnerability, SDO.S
   ],
   relations: [],
   relationsRefs: [objectOrganization],
-  representative: (stix: SDO.StixVulnerability) => {
+  representative: (stix: StixVulnerability) => {
     return stix.name;
   },
   converter_2_1: convertVulnerabilityToStix,
