+++ opencti-platform/opencti-graphql/src/types/stix-2-0-sdo.d.ts
@@ -90,62 +90,6 @@ export interface StixTool extends StixDomainObject {
   tool_version: string;
 }
 
-export interface StixVulnerability extends StixDomainObject {
-  name: string;
-  description: string; // optional
-  x_opencti_cisa_kev: boolean;
-  x_opencti_first_seen_active: Date;
-  // CVSS3
-  x_opencti_cvss_vector_string: string;
-  x_opencti_cvss_base_score: number;
-  x_opencti_cvss_base_severity: string;
-  x_opencti_cvss_attack_vector: string;
-  x_opencti_cvss_attack_complexity: string;
-  x_opencti_cvss_privileges_required: string;
-  x_opencti_cvss_user_interaction: string;
-  x_opencti_cvss_scope: string;
-  x_opencti_cvss_confidentiality_impact: string;
-  x_opencti_cvss_integrity_impact: string;
-  x_opencti_cvss_availability_impact: string;
-  x_opencti_cvss_exploit_code_maturity: string;
-  x_opencti_cvss_remediation_level: string;
-  x_opencti_cvss_report_confidence: string;
-  x_opencti_cvss_temporal_score: number;
-  // CVSS2
-  x_opencti_cvss_v2_vector_string: string;
-  x_opencti_cvss_v2_base_score: number;
-  x_opencti_cvss_v2_access_vector: string;
-  x_opencti_cvss_v2_access_complexity: string;
-  x_opencti_cvss_v2_authentication: string;
-  x_opencti_cvss_v2_confidentiality_impact: string;
-  x_opencti_cvss_v2_integrity_impact: string;
-  x_opencti_cvss_v2_availability_impact: string;
-  x_opencti_cvss_v2_exploitability: string;
-  x_opencti_cvss_v2_remediation_level: string;
-  x_opencti_cvss_v2_report_confidence: string;
-  x_opencti_cvss_v2_temporal_score: number;
-  // CVSS4
-  x_opencti_cvss_v4_vector_string: string;
-  x_opencti_cvss_v4_base_score: number;
-  x_opencti_cvss_v4_base_severity: string;
-  x_opencti_cvss_v4_attack_vector: string;
-  x_opencti_cvss_v4_attack_complexity: string;
-  x_opencti_cvss_v4_attack_requirements: string;
-  x_opencti_cvss_v4_privileges_required: string;
-  x_opencti_cvss_v4_user_interaction: string;
-  x_opencti_cvss_v4_confidentiality_impact_v: string;
-  x_opencti_cvss_v4_confidentiality_impact_s: string;
-  x_opencti_cvss_v4_integrity_impact_v: string;
-  x_opencti_cvss_v4_integrity_impact_s: string;
-  x_opencti_cvss_v4_availability_impact_v: string;
-  x_opencti_cvss_v4_availability_impact_s: string;
-  x_opencti_cvss_v4_exploit_maturity: string;
-  // Others
-  x_opencti_score: number;
-  x_opencti_epss_score: number;
-  x_opencti_epss_percentile: number;
-}
-
 // Location Specific Properties
 // name, description, latitude, longitude, precision, region, country, city, street_address, postal_code
 export interface StixLocation extends StixDomainObject {
