+++ opencti-platform/opencti-graphql/src/types/stix-2-1-sdo.d.ts
@@ -277,67 +277,3 @@ export interface StixTool extends StixDomainObject {
   kill_chain_phases: Array<StixKillChainPhase>; // optional
   tool_version: string;
 }
-
-// Vulnerability Specific Properties
-export interface StixVulnerabilityExtension extends StixOpenctiExtension {
-  // CVSS3
-  cvss_vector: string;
-  cvss_base_score: number;
-  cvss_base_severity: string;
-  cvss_attack_vector: string;
-  cvss_attack_complexity: string;
-  cvss_privileges_required: string;
-  cvss_user_interaction: string;
-  cvss_scope: string;
-  cvss_confidentiality_impact: string;
-  cvss_integrity_impact: string;
-  cvss_availability_impact: string;
-  cvss_exploit_code_maturity: string;
-  cvss_remediation_level: string;
-  cvss_report_confidence: string;
-  cvss_temporal_score: number;
-  // CVSS2
-  cvss_v2_vector: string;
-  cvss_v2_base_score: number;
-  cvss_v2_access_vector: string;
-  cvss_v2_access_complexity: string;
-  cvss_v2_authentication: string;
-  cvss_v2_confidentiality_impact: string;
-  cvss_v2_integrity_impact: string;
-  cvss_v2_availability_impact: string;
-  cvss_v2_exploitability: string;
-  cvss_v2_remediation_level: string;
-  cvss_v2_report_confidence: string;
-  cvss_v2_temporal_score: number;
-  // CVSS4
-  cvss_v4_vector: string;
-  cvss_v4_base_score: number;
-  cvss_v4_base_severity: string;
-  cvss_v4_attack_vector: string;
-  cvss_v4_attack_complexity: string;
-  cvss_v4_attack_requirements: string;
-  cvss_v4_privileges_required: string;
-  cvss_v4_user_interaction: string;
-  cvss_v4_confidentiality_impact_v: string;
-  cvss_v4_confidentiality_impact_s: string;
-  cvss_v4_integrity_impact_v: string;
-  cvss_v4_integrity_impact_s: string;
-  cvss_v4_availability_impact_v: string;
-  cvss_v4_availability_impact_s: string;
-  cvss_v4_exploit_maturity: string;
-  // Others
-  cwe: Array<string>;
-  cisa_kev: boolean;
-  score: number;
-  epss_score: number;
-  epss_percentile: number;
-  first_seen_active: StixDate;
-}
-// name, description
-export interface StixVulnerability extends StixDomainObject {
-  name: string;
-  description: string; // optional
-  extensions: {
-    [STIX_EXT_OCTI]: StixVulnerabilityExtension;
-  };
-}
