+++ opencti-platform/opencti-graphql/tests/03-integration/02-resolvers/vulnerability-test.js
@@ -225,3 +225,156 @@ describe('Vulnerability resolver standard behavior', () => {
     expect(queryResult.data.vulnerability).toBeNull();
   });
 });
+
+// SSVC attributes are gated behind the SSVC_ATTRIBUTES feature flag,
+// which is enabled in the test configuration through `enabled_dev_features: ["*"]`.
+describe('Vulnerability resolver SSVC attributes behavior', () => {
+  let vulnerabilityInternalId;
+  const vulnerabilityStixId = 'vulnerability--3f96c0a1-7dc1-4a35-bd7c-0e2b8c1c9c11';
+
+  const SSVC_READ_QUERY = gql`
+    query vulnerability($id: String!) {
+      vulnerability(id: $id) {
+        id
+        x_opencti_ssvc_exploitation
+        x_opencti_ssvc_automatable
+        x_opencti_ssvc_technical_impact
+      }
+    }
+  `;
+
+  const SSVC_LIST_QUERY = gql`
+    query vulnerabilities($filters: FilterGroup) {
+      vulnerabilities(filters: $filters) {
+        edges {
+          node {
+            id
+            x_opencti_ssvc_exploitation
+          }
+        }
+      }
+    }
+  `;
+
+  it('should create a vulnerability with SSVC attributes', async () => {
+    const CREATE_QUERY = gql`
+      mutation VulnerabilityAdd($input: VulnerabilityAddInput!) {
+        vulnerabilityAdd(input: $input) {
+          id
+          name
+          x_opencti_ssvc_exploitation
+          x_opencti_ssvc_automatable
+          x_opencti_ssvc_technical_impact
+        }
+      }
+    `;
+    const queryResult = await queryAsAdmin({
+      query: CREATE_QUERY,
+      variables: {
+        input: {
+          name: 'Vulnerability with SSVC',
+          stix_id: vulnerabilityStixId,
+          x_opencti_ssvc_exploitation: 'poc',
+          x_opencti_ssvc_automatable: 'no',
+          x_opencti_ssvc_technical_impact: 'partial',
+        },
+      },
+    });
+    const created = queryResult.data.vulnerabilityAdd;
+    expect(created).not.toBeNull();
+    expect(created.x_opencti_ssvc_exploitation).toEqual('poc');
+    expect(created.x_opencti_ssvc_automatable).toEqual('no');
+    expect(created.x_opencti_ssvc_technical_impact).toEqual('partial');
+    vulnerabilityInternalId = created.id;
+  });
+
+  it('should load SSVC attributes on an existing vulnerability', async () => {
+    const queryResult = await queryAsAdmin({ query: SSVC_READ_QUERY, variables: { id: vulnerabilityInternalId } });
+    expect(queryResult.errors).toBeUndefined();
+    expect(queryResult.data.vulnerability.x_opencti_ssvc_exploitation).toEqual('poc');
+    expect(queryResult.data.vulnerability.x_opencti_ssvc_automatable).toEqual('no');
+    expect(queryResult.data.vulnerability.x_opencti_ssvc_technical_impact).toEqual('partial');
+  });
+
+  it('should update SSVC attributes through field patch', async () => {
+    const UPDATE_QUERY = gql`
+      mutation VulnerabilityEdit($id: ID!, $input: [EditInput]!) {
+        vulnerabilityEdit(id: $id) {
+          fieldPatch(input: $input) {
+            id
+            x_opencti_ssvc_exploitation
+            x_opencti_ssvc_automatable
+            x_opencti_ssvc_technical_impact
+          }
+        }
+      }
+    `;
+    const queryResult = await queryAsAdmin({
+      query: UPDATE_QUERY,
+      variables: {
+        id: vulnerabilityInternalId,
+        input: [
+          { key: 'x_opencti_ssvc_exploitation', value: ['active'] },
+          { key: 'x_opencti_ssvc_automatable', value: ['yes'] },
+          { key: 'x_opencti_ssvc_technical_impact', value: ['total'] },
+        ],
+      },
+    });
+    const patched = queryResult.data.vulnerabilityEdit.fieldPatch;
+    expect(patched.x_opencti_ssvc_exploitation).toEqual('active');
+    expect(patched.x_opencti_ssvc_automatable).toEqual('yes');
+    expect(patched.x_opencti_ssvc_technical_impact).toEqual('total');
+  });
+
+  it('should filter vulnerabilities on a SSVC attribute', async () => {
+    const queryResult = await queryAsAdmin({
+      query: SSVC_LIST_QUERY,
+      variables: {
+        filters: {
+          mode: 'and',
+          filters: [{ key: 'x_opencti_ssvc_exploitation', values: ['active'], operator: 'eq', mode: 'or' }],
+          filterGroups: [],
+        },
+      },
+    });
+    expect(queryResult.errors).toBeUndefined();
+    const { edges } = queryResult.data.vulnerabilities;
+    expect(edges.length).toEqual(1);
+    expect(edges[0].node.id).toEqual(vulnerabilityInternalId);
+    expect(edges[0].node.x_opencti_ssvc_exploitation).toEqual('active');
+  });
+
+  it('should reset SSVC attributes when patched with an empty value', async () => {
+    const UPDATE_QUERY = gql`
+      mutation VulnerabilityEdit($id: ID!, $input: [EditInput]!) {
+        vulnerabilityEdit(id: $id) {
+          fieldPatch(input: $input) {
+            id
+            x_opencti_ssvc_exploitation
+          }
+        }
+      }
+    `;
+    const queryResult = await queryAsAdmin({
+      query: UPDATE_QUERY,
+      variables: {
+        id: vulnerabilityInternalId,
+        input: [{ key: 'x_opencti_ssvc_exploitation', value: [null] }],
+      },
+    });
+    expect(queryResult.data.vulnerabilityEdit.fieldPatch.x_opencti_ssvc_exploitation).toBeNull();
+  });
+
+  it('should delete the SSVC vulnerability', async () => {
+    const DELETE_QUERY = gql`
+      mutation vulnerabilityDelete($id: ID!) {
+        vulnerabilityEdit(id: $id) {
+          delete
+        }
+      }
+    `;
+    await queryAsAdmin({ query: DELETE_QUERY, variables: { id: vulnerabilityInternalId } });
+    const queryResult = await queryAsAdmin({ query: SSVC_READ_QUERY, variables: { id: vulnerabilityStixId } });
+    expect(queryResult.data.vulnerability).toBeNull();
+  });
+});
